Refactor message poll models
