--- epp-client/src/epp/request/message.rs
@@ -1,3 +1,0 @@
-//! Types for EPP message requests
-
-pub mod poll;
