--- epp-client/src/epp/response/message.rs
@@ -1,3 +1,0 @@
-//! Types for EPP message responses
-
-pub mod poll;
