+++ epp-client/src/tests/de.rs
@@ -33,6 +33,7 @@ mod response {
     use crate::host::info::EppHostInfoResponse;
     use crate::host::update::EppHostUpdateResponse;
     use crate::message::ack::EppMessageAckResponse;
+    use crate::message::poll::EppMessagePollResponse;
 
     const SVTRID: &str = "RO-6879-1627224678242975";
     const SUCCESS_MSG: &str = "Command completed successfully";
