+++ epp-client/src/tests/se.rs
@@ -38,6 +38,7 @@ mod request {
     use crate::host::update::HostAddRemove;
     use crate::host::update::HostChangeInfo;
     use crate::message::ack::EppMessageAck;
+    use crate::message::poll::EppMessagePoll;
     use chrono::{DateTime, NaiveDate};
     use std::str::FromStr;
